+++ main/boofcv-ip/src/main/java/boofcv/alg/filter/convolve/noborder/ImplConvolveMean_MT.java
@@ -317,7 +317,7 @@ public class ImplConvolveMean_MT {
 				int indexIn = input.startIndex + (y0-radius)*input.stride + x;
 				int indexOut = output.startIndex + output.stride*y0 + x;
 
-				int total = 0;
+				float total = 0;
 				int indexEnd = indexIn + input.stride*kernelWidth;
 				for( ; indexIn < indexEnd; indexIn += input.stride) {
 					total += input.data[indexIn] ;
@@ -391,7 +391,7 @@ public class ImplConvolveMean_MT {
 				int indexIn = input.startIndex + (y0-radius)*input.stride + x;
 				int indexOut = output.startIndex + output.stride*y0 + x;
 
-				int total = 0;
+				double total = 0;
 				int indexEnd = indexIn + input.stride*kernelWidth;
 				for( ; indexIn < indexEnd; indexIn += input.stride) {
 					total += input.data[indexIn] ;
